[FOLLOWERS/ING] Navigation to profile of other users 1. Navigation top row change + display user profile name 2. Added message icon
